+++ backend/libraries/types/src/cryptocurrency.rs
@@ -198,6 +198,19 @@ impl PendingCryptoTransaction {
             }
         }
     }
+
+    pub fn set_memo(mut self, memo: &[u8]) -> Self {
+        match &mut self {
+            PendingCryptoTransaction::NNS(t) => {
+                t.memo = Some(Memo(u64_from_bytes(memo)));
+            }
+            PendingCryptoTransaction::ICRC1(t) => {
+                assert!(memo.len() <= 32);
+                t.memo = Some(memo.to_vec().into());
+            }
+        }
+        self
+    }
 }
 
 impl CompletedCryptoTransaction {
@@ -451,6 +464,18 @@ pub mod icrc1 {
         pub created: TimestampNanos,
         pub error_message: String,
     }
+
+    impl From<CompletedCryptoTransaction> for super::CompletedCryptoTransaction {
+        fn from(value: CompletedCryptoTransaction) -> Self {
+            super::CompletedCryptoTransaction::ICRC1(value)
+        }
+    }
+
+    impl From<FailedCryptoTransaction> for super::FailedCryptoTransaction {
+        fn from(value: FailedCryptoTransaction) -> Self {
+            super::FailedCryptoTransaction::ICRC1(value)
+        }
+    }
 }
 
 impl From<icrc1::PendingCryptoTransaction> for nns::PendingCryptoTransaction {
@@ -464,8 +489,15 @@ impl From<icrc1::PendingCryptoTransaction> for nns::PendingCryptoTransaction {
                 &Subaccount(value.to.subaccount.unwrap_or_default()),
             )),
             fee: Some(Tokens::from_e8s(value.fee.try_into().unwrap())),
-            memo: value.memo.map(|m| Memo(u64::from_be_bytes(m.0.as_ref().try_into().unwrap()))),
+            memo: value.memo.map(|m| Memo(u64_from_bytes(m.0.as_slice()))),
             created: value.created,
         }
     }
+}
+
+fn u64_from_bytes(bytes: &[u8]) -> u64 {
+    assert!(bytes.len() <= 8);
+    let mut u64_bytes = [0u8; 8];
+    u64_bytes[(8 - bytes.len())..].copy_from_slice(bytes);
+    u64::from_be_bytes(u64_bytes)
 }