+++ frontend/app/src/components/home/CurrentChatMenu.svelte
@@ -31,7 +31,6 @@
     import HeartMinus from "../icons/HeartMinus.svelte";
     import HeartPlus from "../icons/HeartPlus.svelte";
     import { interpolateLevel } from "../../utils/i18n";
-    import { OC_GOVERNANCE_CANISTER_ID } from "../../utils/sns";
 
     const client = getContext<OpenChat>("client");
     const dispatch = createEventDispatcher();
@@ -45,10 +44,13 @@
     $: favouritesStore = client.favouritesStore;
     $: messagesRead = client.messagesRead;
     $: isProposalGroup = client.isProposalGroup;
-    $: isChatProposalsGroup =
+    $: governanceCanisterId =
         selectedChatSummary.kind !== "direct_chat" &&
-        selectedChatSummary.subtype?.kind === "governance_proposals" &&
-        selectedChatSummary.subtype.governanceCanisterId === OC_GOVERNANCE_CANISTER_ID;
+        selectedChatSummary.subtype?.kind === "governance_proposals"
+            ? selectedChatSummary.subtype.governanceCanisterId
+            : undefined;
+    $: canMakeProposals =
+        client.tryGetNervousSystem(governanceCanisterId)?.submittingProposalsEnabled ?? false;
     $: userId = selectedChatSummary.kind === "direct_chat" ? selectedChatSummary.them.userId : "";
     $: userStore = client.userStore;
     $: isBot = $userStore[userId]?.kind === "bot";
@@ -369,7 +371,7 @@
                         {/if}
                     {/if}
 
-                    {#if isChatProposalsGroup}
+                    {#if canMakeProposals}
                         <MenuItem on:click={makeProposal}>
                             <ChatQuestionIcon
                                 size={$iconSize}