+++ frontend/openchat-agent/src/services/user/candid/types.d.ts
@@ -1331,6 +1331,7 @@ export interface PrizeContent {
   'prizes_remaining' : number,
   'prizes_pending' : number,
   'caption' : [] | [string],
+  'diamond_only' : boolean,
   'winners' : Array<UserId>,
 }
 export interface PrizeContentInitial {
@@ -1637,12 +1638,16 @@ export interface SnsProposal {
   'proposer' : SnsNeuronId,
 }
 export interface SubmitProposalArgs {
+  'token' : Cryptocurrency,
+  'transaction_fee' : bigint,
+  'ledger' : CanisterId,
   'governance_canister_id' : CanisterId,
   'proposal' : ProposalToSubmit,
+  'proposal_rejection_fee' : bigint,
 }
 export type SubmitProposalResponse = { 'Retrying' : string } |
+  { 'InsufficientPayment' : bigint } |
   { 'Success' : null } |
-  { 'Unauthorized' : null } |
   { 'UserSuspended' : null } |
   { 'GovernanceCanisterNotSupported' : null } |
   { 'TransferFailed' : string } |