+++ frontend/openchat-shared/src/domain/crypto.ts
@@ -21,7 +21,6 @@ export type CryptocurrencyDetails = {
     howToBuyUrl: string;
     infoUrl: string;
     transactionUrlFormat: string;
-    nervousSystem: NervousSystemSummary | undefined;
     added: bigint;
     lastUpdated: bigint;
 };
@@ -33,6 +32,15 @@ export type NervousSystemSummary = {
     isNns: boolean;
     proposalRejectionFee: bigint;
     submittingProposalsEnabled: boolean;
+};
+
+export type NervousSystemDetails = {
+    governanceCanisterId: string;
+    rootCanisterId: string;
+    isNns: boolean;
+    proposalRejectionFee: bigint;
+    submittingProposalsEnabled: boolean;
+    token: CryptocurrencyDetails;
 };
 
 // approximate dollar exchange rates - until we come up with something better
@@ -49,9 +57,9 @@ export const dollarExchangeRates: Record<string, number> = {
     mod: to2SigFigs(dollarToICP / 0.0065),
     cat: to2SigFigs(dollarToICP / 0.0068998),
     boom: to2SigFigs(dollarToICP / 0.00339),
-    icx: to2SigFigs(dollarToICP / 0.0000001),
+    icx: to2SigFigs(dollarToICP / 0.0055),
     nua: to2SigFigs(dollarToICP / 0.0134),
-    sonic: to2SigFigs(dollarToICP / 0.0530),
+    sonic: to2SigFigs(dollarToICP / 0.053),
 };
 
 function to2SigFigs(num: number): number {