+++ frontend/openchat-shared/src/domain/user/user.ts
@@ -1,5 +1,12 @@
 import type { DataContent } from "../data/data";
-import type { Failure, InternalError, NotAuthorised, Retrying, Success, TransferFailed, UserSuspended } from "../response";
+import type {
+    Failure,
+    InternalError,
+    Retrying,
+    Success,
+    TransferFailed,
+    UserSuspended,
+} from "../response";
 
 export type UserOrUserGroup = UserSummary | UserGroupSummary | MentionEveryone;
 
@@ -265,15 +272,19 @@ export type NamedAccount = {
 
 export type SaveCryptoAccountResponse = { kind: "name_taken" } | Success | Failure;
 
-export type SubmitProposalResponse = 
-    Success | 
-    Retrying | 
-    NotAuthorised | 
-    UserSuspended | 
-    GovernanceCanisterNotSupported | 
-    TransferFailed | 
-    InternalError;
+export type SubmitProposalResponse =
+    | Success
+    | Retrying
+    | UserSuspended
+    | GovernanceCanisterNotSupported
+    | InsufficientPayment
+    | TransferFailed
+    | InternalError;
 
 export type GovernanceCanisterNotSupported = {
     kind: "governance_canister_not_supported";
+};
+
+export type InsufficientPayment = {
+    kind: "insufficient_payment";
 };