+++ backend/canisters/market_maker/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Adjust `MemoryManager` bucket size ([#4601](https://github.com/open-chat-labs/open-chat/pull/4601))
+
 ## [[2.0.862](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.862-market_maker)] - 2023-09-26
 
 ### Changed