+++ backend/canisters/notifications_index/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Adjust `MemoryManager` bucket size ([#4601](https://github.com/open-chat-labs/open-chat/pull/4601))
+
 ## [[2.0.794](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.794-notifications_index)] - 2023-08-08
 
 ### Changed